feat(user): implement user module with entity, repository, service, controller and DTOs

+++ src/main/java/com/ai/learning/backend/dto/response/SessionDetailResponse.java
@@ -11,7 +11,7 @@ import lombok.experimental.FieldDefaults;
 @NoArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class SessionDetailResponse {
-    long learningSessionId;
+    Integer learningSessionId;
     String title;
     String videoUrl;
     String transcript;
